+++ linchpin/provision/module_utils/linchpin_rundb.py
@@ -1,11 +1,112 @@
 from __future__ import absolute_import
+from abc import ABCMeta, abstractmethod
+import six
 from tinydb import TinyDB
 from tinydb.storages import JSONStorage
 from tinydb.operations import add
 from tinydb.operations import set as tinySet
 from tinydb.middlewares import CachingMiddleware
-from .basedb import BaseDB
 from six.moves import range
+
+
+class RunDB(six.with_metaclass(ABCMeta, object)):
+
+    @abstractmethod
+    def init_table(self, table):
+        return self.driver.init_table(table)
+
+    @abstractmethod
+    def update_record(self, table, run_id, key, value):
+        return self.driver.update_record(table, run_id, key, value)
+
+    @abstractmethod
+    def get_tx_record(self, tx_id):
+        return self.driver.get_tx_record(tx_id)
+
+    def get_tx_records(self, tx_ids):
+        return self.driver.get_tx_records(tx_ids)
+
+    @abstractmethod
+    def get_record(self, table, action='up', run_id=None):
+        return self.driver.get_record(table, run_id=run_id)
+
+    @abstractmethod
+    def get_records(self, table=[], count=10):
+        return self.driver.get_records(table=table, count=count)
+
+    @abstractmethod
+    def get_tables(self):
+        return self.driver.get_tables()
+
+    @abstractmethod
+    def remove_record(self, table, key, value):
+        return self.driver.remove_record(table, key, value)
+
+    @abstractmethod
+    def search(self, table, key):
+        return self.driver.search(table, key)
+
+    @abstractmethod
+    def query(self, table, query_info):
+        return self.driver.query(table, query_info)
+
+    @abstractmethod
+    def purge(self, table=None):
+        return self.driver.purge(table)
+
+
+class BaseDB(RunDB):
+
+    def __init__(self, driver, conn_str):
+
+        self.name = 'BaseDB'
+        self.conn_str = conn_str
+        self.driver = driver(conn_str)
+
+    def __str__(self):
+        return self.driver.__str__()
+
+    @property
+    def schema(self):
+        return self.driver.schema
+
+    @schema.setter
+    def schema(self, schema):
+        self.driver.schema = schema
+
+    def init_table(self, table):
+        return self.driver.init_table(table)
+
+    def update_record(self, table, run_id, key, value):
+        return self.driver.update_record(table, run_id, key, value)
+
+    def get_tx_record(self, tx_id):
+        return self.driver.get_tx_record(tx_id)
+
+    def get_tx_records(self, tx_ids):
+        return self.driver.get_tx_records(tx_ids)
+
+    def get_record(self, table, action='up', run_id=None):
+        return self.driver.get_record(table, run_id=run_id)
+
+    def get_records(self, table=[], count=10):
+        return self.driver.get_records(table=table, count=count)
+
+    def get_tables(self):
+        return self.driver.get_tables()
+
+    def remove_record(self, table, key, value):
+        return self.driver.remove_record(table, key, value)
+
+    def search(self, table, key):
+        return self.driver.search(table, key)
+
+    def query(self, table, query_info):
+        return self.driver.query(table, query_info)
+
+    def purge(self, table=None):
+        return self.driver.purge(table)
+
 
 
 def usedb(func):
@@ -167,3 +268,12 @@ class TinyRunDB(BaseDB):
 
     def _closedb(self):
         self.db.close()
+
+
+DB_DRIVERS = {
+    "TinyRunDB": TinyRunDB,
+}
+
+
+def get_all_drivers():
+    return DB_DRIVERS
