--- linchpin/rundb/drivers.py
@@ -1,10 +1,0 @@
-from .tinyrundb import TinyRunDB
-
-
-DB_DRIVERS = {
-    "TinyRunDB": TinyRunDB,
-}
-
-
-def get_all_drivers():
-    return DB_DRIVERS
